backend/requirements.txt
django
psycopg2
djangorestframework
markdown
fundoshi==0.0.7
